Simplify report generating
